- added an option to remove_dir(), so it remove only it's content
git-svn-id: http://svn.dnalc.org/ga/pipeline@286 8e20c4c2-b435-4367-b5b9-847ebbf7797d

+++ lib/perl/DNALC/Pipeline/App/Utils.pm
@@ -225,7 +225,7 @@ sub format_datetime {
 }
 
 sub remove_dir {
-	my ($class, $dir) = @_;
+	my ($class, $dir, $keep_root) = @_;
 
 	if (-d $dir) {
 		my @dir_content = <$dir/*>;
@@ -237,7 +237,7 @@ sub remove_dir {
 				unlink $of;
 			} 
 		}
-		rmdir $dir;
+		rmdir $dir unless $keep_root;
 	}
 }
